* Make it compile on freebsd; * Add an 'offset' field (defaulting to zero) that pads the beginning   of each allocation.
The offset field is designed to make the beginning of each malloc'ed
buffer not be page aligned, and thus experiment with cache line
aliasing effects for this benchmark.

The offset isnt' added to each malloc - the offset is added to a static,
global offset that is then added to each malloc.  Ie, if the offset is
32 - the first allocation has the buffer beginning start at 0,
the second starts at 32 bytes into the allocation, the third starts
at 64 bytes into the allocation, etc.